refactor(plugin): split manifest into rdlp-plugin-manifest leaf crate
The example-sign-plugin tool depended on rdlp-plugin just to use
manifest::{canonical_bytes, parse_manifest_str}, which dragged in
wasmtime + sigstore + sled transitively — ~6800 lines of Cargo.lock
for a tool that only signs TOML + WASM bytes.

Lift the manifest types into a new leaf crate, rdlp-plugin-manifest:
  - Manifest, Signature, ManifestError
  - parse_manifest_str / parse_manifest_file
  - validate_plugin_name
  - canonical_bytes
  - KNOWN_CAPABILITIES
Deps: serde + toml + thiserror + sha2 + hex. Nothing else.

rdlp-plugin's manifest module shrinks to a re-export plus a
From<ManifestError> for PluginError bridge, so all existing call
sites compile unchanged. Public API is preserved — including
PluginError::{InvalidManifest, InvalidPluginName, Toml, Io}.

example-sign-plugin's Cargo.lock: 7010 → 463 lines (-93.4%).
Faster cold-build, smaller supply-chain surface for plugin authors.

Verified:
  - cargo check --workspace --all-targets
  - cargo clippy --workspace --all-targets -- -D warnings
  - cargo test -p rdlp-plugin (all integration tests pass)
  - example-sign-plugin compiles standalone after lock regeneration

Closes the W4 sign-plugin Cargo.lock follow-up from the audit.
